chore(modes): Rename "jobs mode" as "cmd mode"
