+++ mode/cmd/doc.go
@@ -1,0 +1,2 @@
+// Package cmd contains the cmd mode implementations and config
+package cmd
