--- mode/jobs/doc.go
@@ -1,2 +1,0 @@
-// Package jobs contains the jobs mode implementations and config
-package jobs
